move ast definition to core module
