+++ .gitignore
@@ -29,3 +29,4 @@ build/
 
 # git mergetool artifact
 *.orig
+gen
