+++ common/build.gradle
@@ -9,7 +9,7 @@ repositories {
 
 dependencies {
     compile "org.antlr:antlr4-runtime:4.7.1"
-    compile group: 'com.google.guava', name: 'guava', version:'15.0'
+    compile group: 'com.google.guava', name: 'guava', version:'23.0'
 
     testCompile group: 'junit', name: 'junit', version: '4.12'
 }
