+++ core/build.gradle
@@ -46,5 +46,11 @@ jacocoTestCoverageVerification {
 
         }
     }
+    afterEvaluate {
+        classDirectories.setFrom(files(classDirectories.files.collect {
+            fileTree(dir: it,
+                    exclude: ['**/ast/**'])
+        }))
+    }
 }
 check.dependsOn jacocoTestCoverageVerification
