+++ core/src/main/java/com/amazon/opendistroforelasticsearch/sql/ast/expression/EqualTo.java
@@ -13,26 +13,27 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression;
+package com.amazon.opendistroforelasticsearch.sql.ast.expression;
 
-import com.amazon.opendistroforelasticsearch.sql.ppl.ast.AbstractNodeVisitor;
+import com.amazon.opendistroforelasticsearch.sql.ast.AbstractNodeVisitor;
 import java.util.Arrays;
 import java.util.List;
+import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 import lombok.ToString;
 
 /**
- * Expression node of logic AND
+ * Expression node of binary operator or comparison relation EQUAL
  */
-@Getter
 @ToString
 @EqualsAndHashCode(callSuper = false)
-@RequiredArgsConstructor
-public class And extends Expression {
-    private final Expression left;
-    private final Expression right;
+@AllArgsConstructor
+public class EqualTo extends Expression {
+    @Getter
+    private Expression left;
+    @Getter
+    private Expression right;
 
     @Override
     public List<Expression> getChild() {
@@ -41,6 +42,6 @@ public class And extends Expression {
 
     @Override
     public <R, C> R accept(AbstractNodeVisitor<R, C> nodeVisitor, C context) {
-        return nodeVisitor.visitAnd(this, context);
+        return nodeVisitor.visitEqualTo(this, context);
     }
 }
