+++ core/src/main/java/com/amazon/opendistroforelasticsearch/sql/ast/expression/Literal.java
@@ -13,35 +13,36 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression;
+package com.amazon.opendistroforelasticsearch.sql.ast.expression;
 
-import com.amazon.opendistroforelasticsearch.sql.ppl.ast.AbstractNodeVisitor;
-import java.util.Arrays;
+import com.amazon.opendistroforelasticsearch.sql.ast.AbstractNodeVisitor;
+import com.google.common.collect.ImmutableList;
 import java.util.List;
-import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import lombok.ToString;
 
 /**
- * Expression node of binary operator or comparison relation EQUAL
+ * Expression node of literal type
+ * Params include literal value (@value) and literal data type (@type) which can be selected from {@link DataType}
  */
+@Getter
 @ToString
 @EqualsAndHashCode(callSuper = false)
-@AllArgsConstructor
-public class EqualTo extends Expression {
-    @Getter
-    private Expression left;
-    @Getter
-    private Expression right;
+@RequiredArgsConstructor
+public class Literal extends Expression {
+
+    private final Object value;
+    private final DataType type;
 
     @Override
     public List<Expression> getChild() {
-        return Arrays.asList(left, right);
+        return ImmutableList.of();
     }
 
     @Override
     public <R, C> R accept(AbstractNodeVisitor<R, C> nodeVisitor, C context) {
-        return nodeVisitor.visitEqualTo(this, context);
+        return nodeVisitor.visitLiteral(this, context);
     }
 }
