+++ core/src/main/java/com/amazon/opendistroforelasticsearch/sql/ast/expression/Or.java
@@ -13,9 +13,9 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression;
+package com.amazon.opendistroforelasticsearch.sql.ast.expression;
 
-import com.amazon.opendistroforelasticsearch.sql.ppl.ast.AbstractNodeVisitor;
+import com.amazon.opendistroforelasticsearch.sql.ast.AbstractNodeVisitor;
 import java.util.Arrays;
 import java.util.List;
 import lombok.EqualsAndHashCode;
@@ -24,22 +24,23 @@ import lombok.RequiredArgsConstructor;
 import lombok.ToString;
 
 /**
- * Expression node of the logic NOT
+ * Expression node of the logic OR
  */
 @Getter
 @ToString
 @EqualsAndHashCode(callSuper = false)
 @RequiredArgsConstructor
-public class Not extends Expression {
-    private final Expression expression;
+public class Or extends Expression {
+    private final Expression left;
+    private final Expression right;
 
     @Override
     public List<Expression> getChild() {
-        return Arrays.asList(expression);
+        return Arrays.asList(left, right);
     }
 
     @Override
     public <R, C> R accept(AbstractNodeVisitor<R, C> nodeVisitor, C context) {
-        return nodeVisitor.visitNot(this, context);
+        return nodeVisitor.visitOr(this, context);
     }
 }
