+++ core/src/main/java/com/amazon/opendistroforelasticsearch/sql/ast/tree/Aggregation.java
@@ -13,10 +13,10 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.sql.ppl.ast.tree;
+package com.amazon.opendistroforelasticsearch.sql.ast.tree;
 
-import com.amazon.opendistroforelasticsearch.sql.ppl.ast.AbstractNodeVisitor;
-import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.Expression;
+import com.amazon.opendistroforelasticsearch.sql.ast.AbstractNodeVisitor;
+import com.amazon.opendistroforelasticsearch.sql.ast.expression.Expression;
 import com.google.common.collect.ImmutableList;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -26,26 +26,31 @@ import lombok.Setter;
 import lombok.ToString;
 
 /**
- * Logical plan node of Project, the interface for building the list of searching fields
+ * Logical plan node of Aggregation, the interface for building aggregation actions in queries
  */
-@ToString
 @Getter
+@Setter
+@ToString
 @EqualsAndHashCode(callSuper = false)
 @AllArgsConstructor
-public class Project extends UnresolvedPlan {
-    @Setter
-    private List<Expression> projectList;
+public class Aggregation extends UnresolvedPlan {
+    private List<Expression> aggExprList;
+    private List<Expression> sortExprList;
+    private List<Expression> groupExprList;
     private UnresolvedPlan child;
 
-    public Project(List<Expression> projectList) {
-        this.projectList = projectList;
+    public Aggregation(List<Expression> aggExprList, List<Expression> sortExprList, List<Expression> groupExprList) {
+        this.aggExprList = aggExprList;
+        this.sortExprList = sortExprList;
+        this.groupExprList = groupExprList;
     }
 
     @Override
-    public Project attach(UnresolvedPlan child) {
+    public Aggregation attach(UnresolvedPlan child) {
         this.child = child;
         return this;
     }
+
 
     @Override
     public List<UnresolvedPlan> getChild() {
@@ -54,7 +59,6 @@ public class Project extends UnresolvedPlan {
 
     @Override
     public <T, C> T accept(AbstractNodeVisitor<T, C> nodeVisitor, C context) {
-
-        return nodeVisitor.visitProject(this, context);
+        return nodeVisitor.visitAggregation(this, context);
     }
 }
