--- gen/OpenDistroSqlLexer.tokens
@@ -1,410 +1,0 @@
-SPACE=1
-SPEC_SQL_COMMENT=2
-COMMENT_INPUT=3
-LINE_COMMENT=4
-ALL=5
-AND=6
-AS=7
-ASC=8
-BETWEEN=9
-BY=10
-CASE=11
-CAST=12
-CROSS=13
-DATETIME=14
-DELETE=15
-DESC=16
-DESCRIBE=17
-DISTINCT=18
-DOUBLE=19
-ELSE=20
-EXISTS=21
-FALSE=22
-FLOAT=23
-FROM=24
-GROUP=25
-HAVING=26
-IN=27
-INNER=28
-INT=29
-IS=30
-JOIN=31
-LEFT=32
-LIKE=33
-LIMIT=34
-LONG=35
-MATCH=36
-NATURAL=37
-NOT=38
-NULL_LITERAL=39
-ON=40
-OR=41
-ORDER=42
-OUTER=43
-REGEXP=44
-RIGHT=45
-SELECT=46
-SHOW=47
-STRING=48
-THEN=49
-TRUE=50
-UNION=51
-USING=52
-WHEN=53
-WHERE=54
-MISSING=55
-EXCEPT=56
-AVG=57
-COUNT=58
-MAX=59
-MIN=60
-SUM=61
-SUBSTRING=62
-TRIM=63
-YEAR=64
-END=65
-FULL=66
-OFFSET=67
-TABLES=68
-ABS=69
-ACOS=70
-ADD=71
-ASCII=72
-ASIN=73
-ATAN=74
-ATAN2=75
-CBRT=76
-CEIL=77
-CONCAT=78
-CONCAT_WS=79
-COS=80
-COSH=81
-COT=82
-CURDATE=83
-DATE=84
-DATE_FORMAT=85
-DAYOFMONTH=86
-DEGREES=87
-E=88
-EXP=89
-EXPM1=90
-FLOOR=91
-IF=92
-IFNULL=93
-ISNULL=94
-LENGTH=95
-LN=96
-LOCATE=97
-LOG=98
-LOG10=99
-LOG2=100
-LOWER=101
-LTRIM=102
-MAKETIME=103
-MODULUS=104
-MONTH=105
-MONTHNAME=106
-MULTIPLY=107
-NOW=108
-PI=109
-POW=110
-POWER=111
-RADIANS=112
-RAND=113
-REPLACE=114
-RINT=115
-ROUND=116
-RTRIM=117
-SIGN=118
-SIGNUM=119
-SIN=120
-SINH=121
-SQRT=122
-SUBTRACT=123
-TAN=124
-TIMESTAMP=125
-UPPER=126
-D=127
-T=128
-TS=129
-LEFT_BRACE=130
-RIGHT_BRACE=131
-DATE_HISTOGRAM=132
-DAY_OF_MONTH=133
-DAY_OF_YEAR=134
-DAY_OF_WEEK=135
-EXCLUDE=136
-EXTENDED_STATS=137
-FIELD=138
-FILTER=139
-GEO_BOUNDING_BOX=140
-GEO_CELL=141
-GEO_DISTANCE=142
-GEO_DISTANCE_RANGE=143
-GEO_INTERSECTS=144
-GEO_POLYGON=145
-HISTOGRAM=146
-HOUR_OF_DAY=147
-INCLUDE=148
-IN_TERMS=149
-MATCHPHRASE=150
-MATCH_PHRASE=151
-MATCHQUERY=152
-MATCH_QUERY=153
-MINUTE_OF_DAY=154
-MINUTE_OF_HOUR=155
-MONTH_OF_YEAR=156
-MULTIMATCH=157
-MULTI_MATCH=158
-NESTED=159
-PERCENTILES=160
-REGEXP_QUERY=161
-REVERSE_NESTED=162
-QUERY=163
-RANGE=164
-SCORE=165
-SECOND_OF_MINUTE=166
-STATS=167
-TERM=168
-TERMS=169
-TOPHITS=170
-WEEK_OF_YEAR=171
-WILDCARDQUERY=172
-WILDCARD_QUERY=173
-STAR=174
-DIVIDE=175
-MODULE=176
-PLUS=177
-MINUS=178
-DIV=179
-MOD=180
-EQUAL_SYMBOL=181
-GREATER_SYMBOL=182
-LESS_SYMBOL=183
-EXCLAMATION_SYMBOL=184
-BIT_NOT_OP=185
-BIT_OR_OP=186
-BIT_AND_OP=187
-BIT_XOR_OP=188
-DOT=189
-LR_BRACKET=190
-RR_BRACKET=191
-COMMA=192
-SEMI=193
-AT_SIGN=194
-ZERO_DECIMAL=195
-ONE_DECIMAL=196
-TWO_DECIMAL=197
-SINGLE_QUOTE_SYMB=198
-DOUBLE_QUOTE_SYMB=199
-REVERSE_QUOTE_SYMB=200
-COLON_SYMB=201
-START_NATIONAL_STRING_LITERAL=202
-STRING_LITERAL=203
-DECIMAL_LITERAL=204
-HEXADECIMAL_LITERAL=205
-REAL_LITERAL=206
-NULL_SPEC_LITERAL=207
-BIT_STRING=208
-DOT_ID=209
-ID=210
-REVERSE_QUOTE_ID=211
-STRING_USER_NAME=212
-ERROR_RECOGNITION=213
-'ALL'=5
-'AND'=6
-'AS'=7
-'ASC'=8
-'BETWEEN'=9
-'BY'=10
-'CASE'=11
-'CAST'=12
-'CROSS'=13
-'DATETIME'=14
-'DELETE'=15
-'DESC'=16
-'DESCRIBE'=17
-'DISTINCT'=18
-'DOUBLE'=19
-'ELSE'=20
-'EXISTS'=21
-'FALSE'=22
-'FLOAT'=23
-'FROM'=24
-'GROUP'=25
-'HAVING'=26
-'IN'=27
-'INNER'=28
-'INT'=29
-'IS'=30
-'JOIN'=31
-'LEFT'=32
-'LIKE'=33
-'LIMIT'=34
-'LONG'=35
-'MATCH'=36
-'NATURAL'=37
-'NOT'=38
-'NULL'=39
-'ON'=40
-'OR'=41
-'ORDER'=42
-'OUTER'=43
-'REGEXP'=44
-'RIGHT'=45
-'SELECT'=46
-'SHOW'=47
-'STRING'=48
-'THEN'=49
-'TRUE'=50
-'UNION'=51
-'USING'=52
-'WHEN'=53
-'WHERE'=54
-'MISSING'=55
-'MINUS'=56
-'AVG'=57
-'COUNT'=58
-'MAX'=59
-'MIN'=60
-'SUM'=61
-'SUBSTRING'=62
-'TRIM'=63
-'YEAR'=64
-'END'=65
-'FULL'=66
-'OFFSET'=67
-'TABLES'=68
-'ABS'=69
-'ACOS'=70
-'ADD'=71
-'ASCII'=72
-'ASIN'=73
-'ATAN'=74
-'ATAN2'=75
-'CBRT'=76
-'CEIL'=77
-'CONCAT'=78
-'CONCAT_WS'=79
-'COS'=80
-'COSH'=81
-'COT'=82
-'CURDATE'=83
-'DATE'=84
-'DATE_FORMAT'=85
-'DAYOFMONTH'=86
-'DEGREES'=87
-'E'=88
-'EXP'=89
-'EXPM1'=90
-'FLOOR'=91
-'IF'=92
-'IFNULL'=93
-'ISNULL'=94
-'LENGTH'=95
-'LN'=96
-'LOCATE'=97
-'LOG'=98
-'LOG10'=99
-'LOG2'=100
-'LOWER'=101
-'LTRIM'=102
-'MAKETIME'=103
-'MODULUS'=104
-'MONTH'=105
-'MONTHNAME'=106
-'MULTIPLY'=107
-'NOW'=108
-'PI'=109
-'POW'=110
-'POWER'=111
-'RADIANS'=112
-'RAND'=113
-'REPLACE'=114
-'RINT'=115
-'ROUND'=116
-'RTRIM'=117
-'SIGN'=118
-'SIGNUM'=119
-'SIN'=120
-'SINH'=121
-'SQRT'=122
-'SUBTRACT'=123
-'TAN'=124
-'TIMESTAMP'=125
-'UPPER'=126
-'D'=127
-'T'=128
-'TS'=129
-'{'=130
-'}'=131
-'DATE_HISTOGRAM'=132
-'DAY_OF_MONTH'=133
-'DAY_OF_YEAR'=134
-'DAY_OF_WEEK'=135
-'EXCLUDE'=136
-'EXTENDED_STATS'=137
-'FIELD'=138
-'FILTER'=139
-'GEO_BOUNDING_BOX'=140
-'GEO_CELL'=141
-'GEO_DISTANCE'=142
-'GEO_DISTANCE_RANGE'=143
-'GEO_INTERSECTS'=144
-'GEO_POLYGON'=145
-'HISTOGRAM'=146
-'HOUR_OF_DAY'=147
-'INCLUDE'=148
-'IN_TERMS'=149
-'MATCHPHRASE'=150
-'MATCH_PHRASE'=151
-'MATCHQUERY'=152
-'MATCH_QUERY'=153
-'MINUTE_OF_DAY'=154
-'MINUTE_OF_HOUR'=155
-'MONTH_OF_YEAR'=156
-'MULTIMATCH'=157
-'MULTI_MATCH'=158
-'NESTED'=159
-'PERCENTILES'=160
-'REGEXP_QUERY'=161
-'REVERSE_NESTED'=162
-'QUERY'=163
-'RANGE'=164
-'SCORE'=165
-'SECOND_OF_MINUTE'=166
-'STATS'=167
-'TERM'=168
-'TERMS'=169
-'TOPHITS'=170
-'WEEK_OF_YEAR'=171
-'WILDCARDQUERY'=172
-'WILDCARD_QUERY'=173
-'*'=174
-'/'=175
-'%'=176
-'+'=177
-'-'=178
-'DIV'=179
-'MOD'=180
-'='=181
-'>'=182
-'<'=183
-'!'=184
-'~'=185
-'|'=186
-'&'=187
-'^'=188
-'.'=189
-'('=190
-')'=191
-','=192
-';'=193
-'@'=194
-'0'=195
-'1'=196
-'2'=197
-'\''=198
-'"'=199
-'`'=200
-':'=201
