+++ ppl/build.gradle
@@ -25,12 +25,13 @@ dependencies {
     antlr "org.antlr:antlr4:4.7.1"
 
     compile "org.antlr:antlr4-runtime:4.7.1"
-    compile group: 'com.google.guava', name: 'guava', version:'15.0'
+    compile group: 'com.google.guava', name: 'guava', version:'23.0'
     compile group: 'org.elasticsearch', name: 'elasticsearch-x-content', version:"${es_version}"
     compile group: 'org.json', name: 'json', version:'20180813'
     compile group: 'org.springframework', name: 'spring-context', version: '5.2.5.RELEASE'
     compile group: 'org.springframework', name: 'spring-beans', version: '5.2.5.RELEASE'
     compile project(':common')
+    compile project(':core')
 
     testCompile group: 'junit', name: 'junit', version: '4.12'
