--- ppl/src/main/java/com/amazon/opendistroforelasticsearch/sql/ppl/ast/expression/DataType.java
@@ -1,26 +1,0 @@
-/*
- *   Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- *   Licensed under the Apache License, Version 2.0 (the "License").
- *   You may not use this file except in compliance with the License.
- *   A copy of the License is located at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *   or in the "license" file accompanying this file. This file is distributed
- *   on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- *   express or implied. See the License for the specific language governing
- *   permissions and limitations under the License.
- */
-
-package com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression;
-
-public enum DataType {
-    TYPE_ERROR,
-    NULL,
-
-    INTEGER,
-    DOUBLE,
-    STRING,
-    BOOLEAN
-}
