+++ ppl/src/main/java/com/amazon/opendistroforelasticsearch/sql/ppl/parser/AstBuilder.java
@@ -17,13 +17,13 @@ package com.amazon.opendistroforelasticsearch.sql.ppl.parser;
 
 import com.amazon.opendistroforelasticsearch.sql.common.utils.StringUtils;
 import com.amazon.opendistroforelasticsearch.sql.ppl.antlr.parser.OpenDistroPPLParserBaseVisitor;
-import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.Expression;
-import com.amazon.opendistroforelasticsearch.sql.ppl.ast.expression.Map;
-import com.amazon.opendistroforelasticsearch.sql.ppl.ast.tree.Aggregation;
-import com.amazon.opendistroforelasticsearch.sql.ppl.ast.tree.Filter;
-import com.amazon.opendistroforelasticsearch.sql.ppl.ast.tree.UnresolvedPlan;
-import com.amazon.opendistroforelasticsearch.sql.ppl.ast.tree.Project;
-import com.amazon.opendistroforelasticsearch.sql.ppl.ast.tree.Relation;
+import com.amazon.opendistroforelasticsearch.sql.ast.expression.Expression;
+import com.amazon.opendistroforelasticsearch.sql.ast.expression.Map;
+import com.amazon.opendistroforelasticsearch.sql.ast.tree.Aggregation;
+import com.amazon.opendistroforelasticsearch.sql.ast.tree.Filter;
+import com.amazon.opendistroforelasticsearch.sql.ast.tree.UnresolvedPlan;
+import com.amazon.opendistroforelasticsearch.sql.ast.tree.Project;
+import com.amazon.opendistroforelasticsearch.sql.ast.tree.Relation;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
